+++ BGAnimations/ScreenSelectMusic overlay/PerPlayer/DensityGraph.lua
@@ -66,9 +66,6 @@ af[#af+1] = Def.Quad{
 		if ThemePrefs.Get("RainbowMode") then
 			self:diffusealpha(0.9)
 		end
-		if ThemePrefs.Get("VisualStyle") == "Technique" then
-			self:diffusealpha(0.5)
-		end
 	end
 }
 
@@ -231,9 +228,6 @@ af2[#af2+1] = Def.ActorFrame{
 	Def.Quad{
 		InitCommand=function(self)
 			self:diffuse(color("#1e282f")):zoomto(width, height)
-			if ThemePrefs.Get("VisualStyle") == "Technique" then
-				self:diffusealpha(0.5)
-			end
 		end,
 	}
 }