+++ BGAnimations/ScreenSelectMusic overlay/SongDescription/SongDescription.lua
@@ -24,9 +24,6 @@ af[#af+1] = Def.Quad{
 		self:diffuse(color("#1e282f"))
 
 		if ThemePrefs.Get("RainbowMode") then self:diffusealpha(0.9) end
-		if ThemePrefs.Get("VisualStyle") == "Technique" then
-			self:diffusealpha(0.5)
-		end
 	end
 }